+++ docs/versionhistory.rst
@@ -5,10 +5,20 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
 **UNRELEASED**
 
+- Updated ``TaskGroup`` to work with asyncio's eager task factories
+  (`#764 <https://github.com/agronholm/anyio/issues/764>`_)
 - Fixed a misleading ``ValueError`` in the context of DNS failures
   (`#815 <https://github.com/agronholm/anyio/issues/815>`_; PR by @graingert)
+- Added the ``wait_readable()`` and ``wait_writable()`` functions which will accept
+  an object with a ``.fileno()`` method or an integer handle, and deprecated
+  their now obsolete versions (``wait_socket_readable()`` and
+  ``wait_socket_writable()``) (PR by @davidbrochart)
+- Fixed the return type annotations of ``readinto()`` and ``readinto1()`` methods in the
+  ``anyio.AsyncFile`` class
+  (`#825 <https://github.com/agronholm/anyio/issues/825>`_)
 - Ported ``ThreadSelectorEventLoop`` from Tornado to allow
-  ``anyio.wait_socket_readable(sock)`` to work on Windows with a ``ProactorEventLoop``.
+  ``anyio.wait_readable()`` and ``anyio.wait_writable()`` to work on Windows with a
+  ``ProactorEventLoop``.
 
 **4.6.2**